Backened/requirements.txt
Flask==3.0.3
flask-cors==4.0.1
opencv-python==4.9.0.80
deepface==0.0.89
requests==2.32.3
numpy==1.26.4
python-dotenv==1.0.1
redis==5.0.3
tf-keras==2.15.1  # Required for TensorFlow compatibility
tensorflow==2.15.1  # Pinned version that works with DeepFace
protobuf==3.20.3  # Specific version to prevent compatibility issues